lift initWc to context + dynamically select namespaces
