+++ packages/renderer/src/renderer/screens/Import/WalletConnect/Import/index.tsx
@@ -20,12 +20,10 @@ import {
 } from '@fortawesome/free-solid-svg-icons';
 
 /** Temp */
-import UniversalProvider from '@walletconnect/universal-provider';
-import { WalletConnectModal } from '@walletconnect/modal';
 import { useAddresses } from '@app/contexts/import/Addresses';
 import { useConnections } from '@app/contexts/common/Connections';
 import { useState } from 'react';
-import { chainIcon } from '@ren/config/chains';
+import { useWalletConnect } from '@ren/renderer/contexts/import/WalletConnect';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { encodeAddress } from '@polkadot/util-crypto';
 import { ellipsisFn } from '@w3ux/utils';
@@ -36,14 +34,7 @@ import {
   ImportAddressRow,
   InfoCard,
 } from '../../Wrappers';
-import type { AnyData } from 'packages/types/src';
-import type { ImportProps, WcFetchedAddress, WcSelectNetwork } from './types';
-
-const WC_PROJECT_ID = 'ebded8e9ff244ba8b6d173b6c2885d87';
-const WC_RELAY_URL = 'wss://relay.walletconnect.com';
-const WC_POLKADOT_CAIP_ID = '91b171bb158e2d3848fa23a9f1c25182';
-const WC_KUSAMA_CAIP_ID = 'b0a8d493285c2df73290dfb7e61f870f';
-const WC_WESTEND_CAIP_ID = 'e143f23803ac50e8f6f8e62695d1ce9e';
+import type { ImportProps, WcFetchedAddress } from './types';
 
 export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
   const { wcAddresses } = useAddresses();
@@ -53,26 +44,7 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
     useState<number>(0);
 
   const theme = darkMode ? themeVariables.darkTheme : themeVariables.lightThene;
-  const [wcNetworks, setWcNetworks] = useState<WcSelectNetwork[]>([
-    {
-      caipId: WC_POLKADOT_CAIP_ID,
-      ChainIcon: chainIcon('Polkadot'),
-      chainId: 'Polkadot',
-      selected: false,
-    },
-    {
-      caipId: WC_KUSAMA_CAIP_ID,
-      ChainIcon: chainIcon('Kusama'),
-      chainId: 'Kusama',
-      selected: false,
-    },
-    {
-      caipId: WC_WESTEND_CAIP_ID,
-      ChainIcon: chainIcon('Westend'),
-      chainId: 'Westend',
-      selected: false,
-    },
-  ]);
+  const { initWc, wcNetworks, setWcNetworks } = useWalletConnect();
 
   // Temp: Addresses for mock UI.
   const mockAddress = '5DwLjkdKkQV6vp12Hi8773GvWAvvkjaxSeVns91u7mswS4io';
@@ -98,101 +70,6 @@ export const Import = ({ setSection, setShowImportUi }: ImportProps) => {
       selected: false,
     },
   ]);
-
-  // Instantiate a universal provider using the projectId created for your app.
-  const initWc = async () => {
-    try {
-      const provider = await UniversalProvider.init({
-        projectId: WC_PROJECT_ID,
-        relayUrl: WC_RELAY_URL,
-      });
-
-      // Listen for WalletConnect events
-      provider.on('session_create', (session: AnyData) => {
-        console.log('Session created:', session);
-      });
-
-      provider.on('session_update', ({ topic, params }: AnyData) => {
-        console.log('Session updated:', topic, params);
-      });
-
-      provider.on('session_delete', (session: AnyData) => {
-        console.log('Session deleted:', session);
-      });
-
-      provider.on('connect', (event: AnyData) => {
-        console.log('Connected to Wallet:', event);
-      });
-
-      provider.on('disconnect', (event: AnyData) => {
-        console.log('Wallet disconnected:', event);
-      });
-
-      // Example Namespace and Sign Client connect call:
-      // Note: this serves as an example where a dapp requires 3 different chain namespaces
-      // (polkadot, kusama and westend). The supported methods, chains, and events
-      // can all be defined by the dapp team based on the requirements of the dapp.
-      const params = {
-        requiredNamespaces: {
-          polkadot: {
-            // Sign the relevant data (either an unsigned transaction or message) and return the signature.
-            methods: ['polkadot_signTransaction', 'polkadot_signMessage'],
-            chains: [
-              `polkadot:${WC_POLKADOT_CAIP_ID}`,
-              `polkadot:${WC_KUSAMA_CAIP_ID}`,
-              `polkadot:${WC_WESTEND_CAIP_ID}`,
-            ],
-            events: ['chainChanged", "accountsChanged'],
-          },
-        },
-      };
-
-      const { uri, approval } = await provider.client.connect(params);
-
-      // Create a standalone modal using your dapps WalletConnect projectId.
-      const walletConnectModal = new WalletConnectModal({
-        enableExplorer: false,
-        explorerRecommendedWalletIds: 'NONE',
-        explorerExcludedWalletIds: 'ALL',
-        projectId: WC_PROJECT_ID,
-      });
-
-      // Open the modal prompting the user to scan the QR code with their wallet app or copy
-      // the URI from the modal and paste into their wallet app to begin the session creation
-      // process.
-      if (uri) {
-        walletConnectModal.openModal({ uri });
-      }
-
-      // Await session approval from the wallet app.
-      // expiry: timestamp of session expiry
-      // pairingTopic: ID of pairing, attempt to connect if cached.
-      // await client.connect({ pairingTopic, requiredNamespaces });
-      const walletConnectSession = await approval();
-
-      walletConnectModal.closeModal();
-
-      // Get the accounts from the session for use in constructing transactions.
-      const wcAccounts = Object.values(walletConnectSession.namespaces)
-        .map((namespace) => namespace.accounts)
-        .flat();
-
-      // Grab account addresses from CAIP account formatted accounts
-      const accounts = wcAccounts.map((wcAccount) => {
-        const address = wcAccount.split(':')[2];
-        return address;
-      });
-
-      console.log(accounts);
-    } catch (error: AnyData) {
-      if (error.message.includes('Proposal expired')) {
-        console.warn('Proposal expired. Prompting user to retry.');
-        // Optionally retry or guide the user
-      } else {
-        console.error('An unexpected error occurred:', error);
-      }
-    }
-  };
 
   return (
     <Scrollable
